requirements.txt

pandas
numpy
torch==1.10.0
pytorch-lightning==1.5.10
faiss-cpu==1.7.2